fix(events): widen task target_value to BIGINT so multi-billion GP goals save
Creating a loot_value task with a goal past signed-INT max returned a 500.
web_event_tasks.target_value was INT, so "Obtain 5b in Value from All Drops"
(target_value 5,000,000,000) reached MySQL as error 1264 and surfaced as an
unhandled DataError. The optional target NPC was a red herring — the payload
carried target=None fine; only the magnitude failed.

- target_value -> BigInteger on EventTask and EventTaskLibraryItem, matching
  the columns it is folded against (EventCompletion.quantity BigInteger,
  EventProgress.progress DOUBLE).
- _require_target_value now applies MAX_TARGET_VALUE (2**53-1, JS-safe) when a
  kind has no tighter bound, so out-of-range input 422s instead of 500ing.
- Bound task points on the create route (it had no validation at all) and add
  the matching ceiling on the PATCH route; points stays INT, so MAX_TASK_POINTS
  keeps it under signed-INT max.

Migration web69a_task_target_value_bigint (alembic/versions is gitignored):
alembic upgrade web69a_task_target_value_bigint. Already applied in prod.

+++ db/models/events.py
@@ -390,7 +390,11 @@ class EventTask(Base):
     type = Column(String(24), nullable=False)
     label = Column(String(255), nullable=False)
     target = Column(String(120), nullable=True)
-    target_value = Column(Integer, nullable=True)
+    # BigInteger (web69a): loot_value goals are raw GP, and a multi-billion GP
+    # target blew past signed-INT 2.147B — the INSERT died with MySQL 1264
+    # rather than a readable error (same lesson as ``EventCompletion.quantity``,
+    # P1-7, which this is compared against).
+    target_value = Column(BigInteger, nullable=True)
     points = Column(Integer, nullable=False, default=0)
     requires_confirmation = Column(Boolean, nullable=False, default=False)  # per-task flag (PRD D3)
     config = Column(Text, nullable=True)  # JSON: any_of/assembly/point_collection item lists etc.
@@ -845,7 +849,7 @@ class EventTaskLibraryItem(Base):
     description = Column(Text, nullable=True)
     type = Column(String(24), nullable=False)  # EVENT_TASK_TYPES
     target = Column(String(120), nullable=True)
-    target_value = Column(Integer, nullable=True)
+    target_value = Column(BigInteger, nullable=True)  # BigInteger — see EventTask.target_value
     default_points = Column(Integer, nullable=False, default=0)
     difficulty = Column(String(24), nullable=True)  # air|water|earth|fire (legacy tiers)
     config = Column(Text, nullable=True)  # JSON: item lists / semantics
